Reference

​

Contents

​

Reference

Contents

Index

​

Index

​

CalculusTreeTools.abstract_tree.ab_tree

CalculusTreeTools.M_evaluation_expr_tree._evaluate_expr_tree_multiple_points

CalculusTreeTools.M_evaluation_expr_tree.calcul_Hessian_expr_tree

CalculusTreeTools.M_evaluation_expr_tree.calcul_gradient_expr_tree

CalculusTreeTools.abstract_tree.create_tree

CalculusTreeTools.algo_expr_tree.cast_type_of_constant

CalculusTreeTools.algo_expr_tree.delete_imbricated_plus

CalculusTreeTools.algo_expr_tree.element_fun_from_N_to_Ni!

CalculusTreeTools.algo_expr_tree.get_Ui

CalculusTreeTools.algo_expr_tree.get_elemental_variable

CalculusTreeTools.algo_expr_tree.get_type_tree

CalculusTreeTools.bound_propagations.set_bounds!

CalculusTreeTools.calcul_Hessian_expr_tree

CalculusTreeTools.calcul_gradient_expr_tree

CalculusTreeTools.calcul_gradient_expr_tree2

CalculusTreeTools.cast_type_of_constant

CalculusTreeTools.delete_imbricated_plus

CalculusTreeTools.element_fun_from_N_to_Ni!

CalculusTreeTools.evaluate_expr_tree

CalculusTreeTools.get_Ui

CalculusTreeTools.get_elemental_variable

CalculusTreeTools.get_function_of_evaluation

CalculusTreeTools.get_type_tree

CalculusTreeTools.interface_type_expr._type_power

CalculusTreeTools.interface_type_expr._type_product

CalculusTreeTools.trait_expr_tree.expr_tree_to_create

CalculusTreeTools.trait_expr_tree.get_real_node

CalculusTreeTools.trait_expr_tree.transform_to_Expr

CalculusTreeTools.trait_expr_tree.transform_to_expr_tree

CalculusTreeTools.transform_to_Expr

CalculusTreeTools.transform_to_Expr_julia

CalculusTreeTools.transform_to_expr_tree

​

CalculusTreeTools.calcul_Hessian_expr_tree — Method

calcul_Hessian_expr_tree(t, x)

Evaluate the Hessian of the calculus tree t as the point x calculHessianexpr_tree( :(x[1]^2 + x[2]), rand(2))

[2.0 0.0; 0.0 0.0]

CalculusTreeTools.calcul_gradient_expr_tree — Method

calcul_gradient_expr_tree(t, x)

Evaluate the gradient of the calculus tree t as the point x calculgradientexpr_tree( :(x[1] + x[2]), rand(2))

[1.0 1.0]

CalculusTreeTools.calcul_gradient_expr_tree2 — Method

calcul_gradient_expr_tree2(t, x)

Evaluate the gradient of the calculus tree t as the point x calculgradientexpr_tree2( :(x[1] + x[2]), rand(2))

[1.0 1.0]

CalculusTreeTools.cast_type_of_constant — Method

cast_type_of_constant(expr_tree, t)

Cast the constant of the Calculus tree expr_tree to the type t.

CalculusTreeTools.delete_imbricated_plus — Method

delete_imbricated_plus(e)

If e represent a calculus tree, deleteimbricatedplus(e) will split that function into element function if it is possible. concretely if divides the tree into subtrees as long as top nodes are + or -

deleteimbricatedplus(:(x[1] + x[2] + x[3]*x[4] ) ) [ x[1], x[2], x[3] * x[4] ]

CalculusTreeTools.element_fun_from_N_to_Ni! — Method

element_fun_from_N_to_Ni!(expr_tree, v)

Transform the tree exprtree, which represent a function from Rⁿ ⇢ R, to an element, function from Rⁱ → R, where i is the length of the vector v . This function rename the variable of exprtree to x₁,x₂,... instead of x₇,x₉ for example elementfunfromNto_Ni!( :(x[4] + x[5]), [1,2])

:(x[1] + x[2])

CalculusTreeTools.evaluate_expr_tree — Method

evaluate_expr_tree(t, x)

Evaluate the Calculus tree t given the vector x, the value of the variables in t. evaluateexprtree(:(x[1] + x[2]), ones(2))

2

evaluateexprtree(:(x[1] + x[2]), [0,1])

1

CalculusTreeTools.get_Ui — Method

get_Ui(index_new_var, n)

Create a the matrix U associated to the variable appearing in indexnewvar. This function create a sparse matrix of size length(indexnewvar)×n.

CalculusTreeTools.get_elemental_variable — Method

get_elemental_variable(expr_tree)

Return the index of the variable appearing in the expression tree getelementalvariable( :(x[1] + x[3]) )

[1, 3]

getelementalvariable( :(x[1]^2 + x[6] + x[2]) )

[1, 6, 2]

CalculusTreeTools.get_function_of_evaluation — Method

get_function_of_evaluation(ex)

Return a evaluation function of ex with better performance than the actual evaluateexprtree.

CalculusTreeTools.get_type_tree — Method

get_type_tree(t)

Return the type of the expression tree t, whose the types are defined in typeexpr/impltype_expr.jl

gettypetree( :(5+4)) = constant gettypetree( :(x[1])) = linear gettypetree( :(x[1]* x[2])) = quadratic

CalculusTreeTools.transform_to_Expr — Method

transform_to_Expr(expr_tree)

Transform into an Expr the parameter exprtree if exprtree satisfies the trait defined in traitexprtree. ATTENTION: This function return an Expr with variable as MathOptInterface.VariableIndex In order to get an standard Expr use transformtoExpr_julia.

CalculusTreeTools.transform_to_Expr_julia — Method

transform_to_Expr_julia(expr_tree)

Transform into an Expr the parameter exprtree if exprtree satisfies the trait define in traitexprtree.

CalculusTreeTools.transform_to_expr_tree — Method

transform_to_expr_tree(Expr)

Transform into an exprtree the parameter Expr if exprtree satisfies the trait define in traitexprtree

CalculusTreeTools.trait_expr_tree.expr_tree_to_create — Method

exprtreeto_create(arbre créé, arbre d'origine) fonction ayant pour but d'homogénéiser 2 arbres quelconque peut importe leurs types.

CalculusTreeTools.trait_expr_tree.get_real_node — Method

getrealnode(a) Fonction à prendre avec des pincettes, pour le moment utiliser seulement sur les feuilles.

CalculusTreeTools.trait_expr_tree.transform_to_Expr — Method

transformtoExpr(ex) This function transform an expr_tree and transform it in Expr. #Deprecated

CalculusTreeTools.trait_expr_tree.transform_to_expr_tree — Method

transformtoexprtree(expressiontree) This function takes an argument expressiontree satisfying the trait isexprtree and return an expression tree of the type texprtree. This function is usefull in our algorithms to synchronise all the types satisfying the trait isexprtree (like Expr) to the type texpr_tree.

CalculusTreeTools.M_evaluation_expr_tree._evaluate_expr_tree_multiple_points — Method

evaluateexprtreemultiplepoints(exprtree, vecx) Evaluate the exprtree as many times as there is vector in vec_x

CalculusTreeTools.M_evaluation_expr_tree.calcul_Hessian_expr_tree — Method

calcul_Hessian_expr_tree(expr_tree, x)

Compute the hessian matrix of the function represented by expr_tree at the point x

CalculusTreeTools.M_evaluation_expr_tree.calcul_gradient_expr_tree — Method

calcul_gradient_expr_tree(expr_tree, x)

Evaluation the gradient of the function represented by expr_tree at the point x

CalculusTreeTools.bound_propagations.set_bounds! — Method

set_bounds!(tree,bound_tre)

Propagate the bounds for each node of tree

CalculusTreeTools.algo_expr_tree.cast_type_of_constant — Method

cast_type_of_constant(expr_tree, t)

Cast the constant of the expression tree expr_tree to the type t.

CalculusTreeTools.algo_expr_tree.delete_imbricated_plus — Method

delete_imbricated_plus(t)

t must be a type which satisfies the trait_expr_tree. In that case if
t represent a function, delete_imbricated_plus(t) will split that function
into element function if it is possible.

delete_imbricated_plus(:(x[1] + x[2] + x[3]*x[4] ) )
[
x[1],
x[2],
x[3] * x[4]
]

CalculusTreeTools.algo_expr_tree.element_fun_from_N_to_Ni! — Method

element_fun_from_N_to_Ni!(expr_tree, vector)

Transform the tree exprtree, which represent a function from Rⁿ ⇢ R, to an element function from Rⁱ → R . This function rename the variable of exprtree to x₁,x₂,... instead of x₇,x₉ for example

CalculusTreeTools.algo_expr_tree.get_Ui — Method

get_Ui(index_new_var, n)

Create a the matrix U associated to the variable appearing in indexnewvar. This function create a sparse matrix of size length(indexnewvar)×n.

CalculusTreeTools.algo_expr_tree.get_elemental_variable — Method

get_elemental_variable(expr_tree)

Return the index of the variable appearing in the expression tree

get_elemental_variable( :(x[1] + x[3]) )
> [1, 3]
get_elemental_variable( :(x[1]^2 + x[6] + x[2]) )
> [1, 6, 2]

CalculusTreeTools.algo_expr_tree.get_type_tree — Method

get_type_tree(t)

Return the type of the expression tree t, whose the type is inside the trait_expr_tree

get_type_tree( :(5+4)) = constant
get_type_tree( :(x[1])) = linear
get_type_tree( :(x[1]* x[2])) = quadratic

CalculusTreeTools.abstract_tree.ab_tree — Type

Define the abstrait type ab_tree

CalculusTreeTools.abstract_tree.create_tree — Method

createtree() Define a tree, the implementation of the function must be done by the subtype of abtree

CalculusTreeTools.interface_type_expr._type_power — Method

_type_product(index,b)

Compute the ttypeexpr_basic that results of b^(index)

CalculusTreeTools.interface_type_expr._type_product — Method

_type_product(a,b)

Compute the ttypeexpr_basic that resultsof of the product a*b