Reference

​

Contents

​

Reference

Contents

Index

​

Index

​

ExpressionTreeForge.Complete_expr_tree

ExpressionTreeForge.M_abstract_expr_node.Abstract_expr_node

ExpressionTreeForge.M_abstract_expr_node.MyRef

ExpressionTreeForge.M_abstract_expr_tree.AbstractExprTree

ExpressionTreeForge.M_abstract_expr_tree.Bounds

ExpressionTreeForge.M_abstract_tree.AbstractTree

ExpressionTreeForge.M_bound_propagations.Bound_tree

ExpressionTreeForge.M_convexity_detection.Convexity_tree

ExpressionTreeForge.M_implementation_complete_expr_tree.Complete_expr_tree

ExpressionTreeForge.M_implementation_complete_expr_tree.Complete_node

ExpressionTreeForge.M_implementation_convexity_type.Convexity_type

ExpressionTreeForge.M_implementation_convexity_type.Convexity_wrapper

ExpressionTreeForge.M_implementation_expr_tree.Type_expr_tree

ExpressionTreeForge.M_implementation_expression_tree_Expr.Variable_counter

ExpressionTreeForge.M_implementation_tree.Type_node

ExpressionTreeForge.M_trait_expr_node.Type_expr_node

ExpressionTreeForge.M_trait_expr_node.Type_not_expr_node

ExpressionTreeForge.M_trait_expr_tree.Is_expr_tree

ExpressionTreeForge.M_trait_expr_tree.Is_not_expr_tree

ExpressionTreeForge.M_trait_tree.Type_not_trait_tree

ExpressionTreeForge.M_trait_tree.Type_trait_tree

ExpressionTreeForge.M_trait_type_expr.Type_not_type_expr

ExpressionTreeForge.M_trait_type_expr.Type_type_expr

ExpressionTreeForge.Type_calculus_tree

ExpressionTreeForge.Type_expr_tree

ExpressionTreeForge.Type_node

ExpressionTreeForge.M_abstract_expr_node.create_new_vector_myRef

ExpressionTreeForge.M_abstract_expr_node.create_node_expr

ExpressionTreeForge.M_abstract_expr_node.create_undef_array_myRef

ExpressionTreeForge.M_abstract_expr_node.create_vector_of_vector_myRef

ExpressionTreeForge.M_abstract_expr_node.equalize_vec_vec_myRef!

ExpressionTreeForge.M_abstract_expr_node.get_myRef

ExpressionTreeForge.M_abstract_expr_node.new_ref

ExpressionTreeForge.M_abstract_expr_node.set_myRef!

ExpressionTreeForge.M_abstract_expr_tree.create_Expr

ExpressionTreeForge.M_abstract_expr_tree.create_Expr2

ExpressionTreeForge.M_abstract_expr_tree.create_empty_bounds

ExpressionTreeForge.M_abstract_expr_tree.create_expr_tree

ExpressionTreeForge.M_abstract_expr_tree.get_bounds

ExpressionTreeForge.M_abstract_expr_tree.set_bound!

ExpressionTreeForge.M_abstract_tree.create_tree

ExpressionTreeForge.M_bound_propagations.create_bounds_tree

ExpressionTreeForge.M_bound_propagations.get_bound

ExpressionTreeForge.M_bound_propagations.set_bounds!

ExpressionTreeForge.M_convexity_detection.concave_type

ExpressionTreeForge.M_convexity_detection.constant_type

ExpressionTreeForge.M_convexity_detection.convex_type

ExpressionTreeForge.M_convexity_detection.create_convex_tree

ExpressionTreeForge.M_convexity_detection.get_convexity_status

ExpressionTreeForge.M_convexity_detection.linear_type

ExpressionTreeForge.M_convexity_detection.set_convexity!

ExpressionTreeForge.M_convexity_detection.unknown_type

ExpressionTreeForge.M_evaluation_expr_tree.gradient_expr_tree_forward

ExpressionTreeForge.M_evaluation_expr_tree.hessian_expr_tree

ExpressionTreeForge.M_implementation_convexity_type.concave_type

ExpressionTreeForge.M_implementation_convexity_type.concave_wrapper

ExpressionTreeForge.M_implementation_convexity_type.constant_type

ExpressionTreeForge.M_implementation_convexity_type.constant_wrapper

ExpressionTreeForge.M_implementation_convexity_type.convex_type

ExpressionTreeForge.M_implementation_convexity_type.convex_wrapper

ExpressionTreeForge.M_implementation_convexity_type.get_convexity_wrapper

ExpressionTreeForge.M_implementation_convexity_type.init_conv_status

ExpressionTreeForge.M_implementation_convexity_type.is_concave

ExpressionTreeForge.M_implementation_convexity_type.is_constant

ExpressionTreeForge.M_implementation_convexity_type.is_convex

ExpressionTreeForge.M_implementation_convexity_type.is_linear

ExpressionTreeForge.M_implementation_convexity_type.is_not_treated

ExpressionTreeForge.M_implementation_convexity_type.is_treated

ExpressionTreeForge.M_implementation_convexity_type.is_unknown

ExpressionTreeForge.M_implementation_convexity_type.linear_type

ExpressionTreeForge.M_implementation_convexity_type.linear_wrapper

ExpressionTreeForge.M_implementation_convexity_type.not_treated_type

ExpressionTreeForge.M_implementation_convexity_type.set_convexity_wrapper!

ExpressionTreeForge.M_implementation_convexity_type.unknown_type

ExpressionTreeForge.M_implementation_convexity_type.unknown_wrapper

ExpressionTreeForge.M_implementation_tree.equal_tree

ExpressionTreeForge.M_implementation_type_expr.return_constant

ExpressionTreeForge.M_implementation_type_expr.return_cubic

ExpressionTreeForge.M_implementation_type_expr.return_linear

ExpressionTreeForge.M_implementation_type_expr.return_more

ExpressionTreeForge.M_implementation_type_expr.return_quadratic

ExpressionTreeForge.M_interface_expr_node._cast_constant!

ExpressionTreeForge.M_interface_expr_node._change_from_N_to_Ni!

ExpressionTreeForge.M_interface_expr_node._evaluate_node

ExpressionTreeForge.M_interface_expr_node._evaluate_node!

ExpressionTreeForge.M_interface_expr_node._get_type_node

ExpressionTreeForge.M_interface_expr_node._get_var_index

ExpressionTreeForge.M_interface_expr_node._node_bound

ExpressionTreeForge.M_interface_expr_node._node_convexity

ExpressionTreeForge.M_interface_expr_node._node_is_constant

ExpressionTreeForge.M_interface_expr_node._node_is_cos

ExpressionTreeForge.M_interface_expr_node._node_is_minus

ExpressionTreeForge.M_interface_expr_node._node_is_operator

ExpressionTreeForge.M_interface_expr_node._node_is_plus

ExpressionTreeForge.M_interface_expr_node._node_is_power

ExpressionTreeForge.M_interface_expr_node._node_is_sin

ExpressionTreeForge.M_interface_expr_node._node_is_tan

ExpressionTreeForge.M_interface_expr_node._node_is_times

ExpressionTreeForge.M_interface_expr_node._node_is_variable

ExpressionTreeForge.M_interface_expr_node._node_to_Expr

ExpressionTreeForge.M_interface_expr_node._node_to_Expr2

ExpressionTreeForge.M_interface_expr_tree._expr_tree_to_create

ExpressionTreeForge.M_interface_expr_tree._get_expr_children

ExpressionTreeForge.M_interface_expr_tree._get_expr_node

ExpressionTreeForge.M_interface_expr_tree._get_real_node

ExpressionTreeForge.M_interface_expr_tree._inverse_expr_tree

ExpressionTreeForge.M_interface_expr_tree._transform_to_expr_tree

ExpressionTreeForge.M_interface_tree._get_children

ExpressionTreeForge.M_interface_tree._get_children

ExpressionTreeForge.M_interface_tree._get_node

ExpressionTreeForge.M_interface_tree._get_node

ExpressionTreeForge.M_interface_type_expr._is_constant

ExpressionTreeForge.M_interface_type_expr._is_constant

ExpressionTreeForge.M_interface_type_expr._is_cubic

ExpressionTreeForge.M_interface_type_expr._is_cubic

ExpressionTreeForge.M_interface_type_expr._is_linear

ExpressionTreeForge.M_interface_type_expr._is_linear

ExpressionTreeForge.M_interface_type_expr._is_more

ExpressionTreeForge.M_interface_type_expr._is_more

ExpressionTreeForge.M_interface_type_expr._is_quadratic

ExpressionTreeForge.M_interface_type_expr._is_quadratic

ExpressionTreeForge.M_interface_type_expr._type_power

ExpressionTreeForge.M_interface_type_expr._type_power

ExpressionTreeForge.M_interface_type_expr._type_product

ExpressionTreeForge.M_interface_type_expr._type_product

ExpressionTreeForge.M_trait_expr_node.cast_constant!

ExpressionTreeForge.M_trait_expr_node.change_from_N_to_Ni!

ExpressionTreeForge.M_trait_expr_node.evaluate_node

ExpressionTreeForge.M_trait_expr_node.evaluate_node!

ExpressionTreeForge.M_trait_expr_node.get_type_node

ExpressionTreeForge.M_trait_expr_node.get_var_index

ExpressionTreeForge.M_trait_expr_node.node_bound

ExpressionTreeForge.M_trait_expr_node.node_convexity

ExpressionTreeForge.M_trait_expr_node.node_is_constant

ExpressionTreeForge.M_trait_expr_node.node_is_cos

ExpressionTreeForge.M_trait_expr_node.node_is_minus

ExpressionTreeForge.M_trait_expr_node.node_is_operator

ExpressionTreeForge.M_trait_expr_node.node_is_plus

ExpressionTreeForge.M_trait_expr_node.node_is_power

ExpressionTreeForge.M_trait_expr_node.node_is_sin

ExpressionTreeForge.M_trait_expr_node.node_is_tan

ExpressionTreeForge.M_trait_expr_node.node_is_times

ExpressionTreeForge.M_trait_expr_node.node_is_variable

ExpressionTreeForge.M_trait_expr_node.node_to_Expr

ExpressionTreeForge.M_trait_expr_node.node_to_Expr2

ExpressionTreeForge.M_trait_expr_tree.expr_tree_equal

ExpressionTreeForge.M_trait_expr_tree.expr_tree_to_create

ExpressionTreeForge.M_trait_expr_tree.get_expr_children

ExpressionTreeForge.M_trait_expr_tree.get_expr_node

ExpressionTreeForge.M_trait_expr_tree.get_real_node

ExpressionTreeForge.M_trait_expr_tree.inverse_expr_tree

ExpressionTreeForge.M_trait_expr_tree.is_expr_tree

ExpressionTreeForge.M_trait_expr_tree.transform_to_Expr

ExpressionTreeForge.M_trait_expr_tree.transform_to_expr_tree

ExpressionTreeForge.M_trait_tree.get_children

ExpressionTreeForge.M_trait_tree.get_node

ExpressionTreeForge.M_trait_tree.is_type_trait_tree

ExpressionTreeForge.M_trait_type_expr.is_constant

ExpressionTreeForge.M_trait_type_expr.is_cubic

ExpressionTreeForge.M_trait_type_expr.is_linear

ExpressionTreeForge.M_trait_type_expr.is_more

ExpressionTreeForge.M_trait_type_expr.is_quadratic

ExpressionTreeForge.M_trait_type_expr.is_trait_type_expr

ExpressionTreeForge.M_trait_type_expr.type_power

ExpressionTreeForge.M_trait_type_expr.type_product

ExpressionTreeForge.algo_expr_tree.N_to_Ni

ExpressionTreeForge.algo_expr_tree.cast_type_of_constant

ExpressionTreeForge.algo_expr_tree.delete_imbricated_plus

ExpressionTreeForge.algo_expr_tree.element_fun_from_N_to_Ni!

ExpressionTreeForge.algo_expr_tree.get_Ui

ExpressionTreeForge.algo_expr_tree.get_elemental_variable

ExpressionTreeForge.algo_expr_tree.get_type_tree

ExpressionTreeForge.algo_tree.printer_tree

ExpressionTreeForge.cast_type_of_constant

ExpressionTreeForge.concave_type

ExpressionTreeForge.constant_type

ExpressionTreeForge.convex_type

ExpressionTreeForge.create_bounds_tree

ExpressionTreeForge.create_complete_tree

ExpressionTreeForge.create_convex_tree

ExpressionTreeForge.delete_imbricated_plus

ExpressionTreeForge.element_fun_from_N_to_Ni!

ExpressionTreeForge.evaluate_expr_tree

ExpressionTreeForge.evaluate_expr_tree_multiple_points

ExpressionTreeForge.get_Ui

ExpressionTreeForge.get_bound

ExpressionTreeForge.get_convexity_status

ExpressionTreeForge.get_elemental_variable

ExpressionTreeForge.get_function_of_evaluation

ExpressionTreeForge.get_type_tree

ExpressionTreeForge.gradient_expr_tree_forward

ExpressionTreeForge.gradient_expr_tree_reverse

ExpressionTreeForge.hessian_expr_tree

ExpressionTreeForge.is_concave

ExpressionTreeForge.is_constant

ExpressionTreeForge.is_constant

ExpressionTreeForge.is_convex

ExpressionTreeForge.is_cubic

ExpressionTreeForge.is_linear

ExpressionTreeForge.is_linear

ExpressionTreeForge.is_more

ExpressionTreeForge.is_not_treated

ExpressionTreeForge.is_quadratic

ExpressionTreeForge.is_treated

ExpressionTreeForge.is_unknown

ExpressionTreeForge.linear_type

ExpressionTreeForge.print_tree

ExpressionTreeForge.set_bounds!

ExpressionTreeForge.set_convexity!

ExpressionTreeForge.transform_to_Expr

ExpressionTreeForge.transform_to_Expr_julia

ExpressionTreeForge.transform_to_expr_tree

ExpressionTreeForge.unknown_type

​

ExpressionTreeForge.Complete_expr_tree — Type

Complete_expr_tree{T} <: AbstractTree

Implementation of an expression tree. Complete_expr_tree is the same than Type_expr_tree with the additions in each node of a Bounds and Convexity_wrapper. A Complete_expr_tree has fields:

field::Complete_node{T} representing an operator, a constant or a variable alongide its bounds and its convexity status;

children::Vector{Complete_expr_tree{T}} a vector of children, each of them being a Complete_expr_tree{T}.

ExpressionTreeForge.Type_calculus_tree — Type

Type_calculus_tree

Represent the different types that a expression can be :

constant

linear

quadratic

cubic

more which means non-linear, non-quadratic and non-cubic.

ExpressionTreeForge.Type_expr_tree — Type

Type_expr_tree{T} <: AbstractTree

Basic implementation of an expression tree. Every expression tree supported must be able to return Type_expr_tree. A Type_expr_tree has fields:

field::Abstract_expr_node representing an operator, a constant or a variable;

children::Vector{Type_expr_tree{T}} a vector of children, each of them being a Type_expr_tree{T}.

ExpressionTreeForge.Type_node — Type

Type_node{T} <: AbstractTree

Basic implementation of a tree. A Type_node has fields:

field gathering the informations about the current node;

children a vector of children, each of them being a Type_node.

ExpressionTreeForge.cast_type_of_constant — Method

cast_type_of_constant(expr_tree, type::DataType)

Cast to type the constants of expr_tree.

ExpressionTreeForge.concave_type — Method

concave = concave_type()

Return the value of concave from the enumerative type M_implementation_convexity_type.Convexity_type.

ExpressionTreeForge.constant_type — Method

constant = constant_type()

Return the value of constant from the enumerative type M_implementation_convexity_type.Convexity_type.

ExpressionTreeForge.convex_type — Method

convex = convex_type()

Return the value of convex from the enumerative type M_implementation_convexity_type.Convexity_type.

ExpressionTreeForge.create_bounds_tree — Method

bound_tree = create_bounds_tree(tree)

Return a similar expression tree to tree, where each node has an undefined bounds.

ExpressionTreeForge.create_complete_tree — Method

complete_tree = create_complete_tree(expression_tree)

Create a complete_tree::Complete_expr_tree from expression_tree (::Type_expr_tree for now).

ExpressionTreeForge.create_convex_tree — Method

convex_tree = create_convex_tree(tree::Type_node)

Return a similar expression tree to tree, where each node has an undefined convexity status.

ExpressionTreeForge.delete_imbricated_plus — Method

separated_terms = delete_imbricated_plus(expr_tree)

Divide the expression tree as a terms of a sum if possible. It returns a vector where each component is a subexpression tree of expr_tree.

Example:

julia> delete_imbricated_plus(:(x[1] + x[2] + x[3]*x[4] ) )
3-element Vector{Expr}:
 :(x[1])
 :(x[2])
 :(x[3] * x[4])

ExpressionTreeForge.element_fun_from_N_to_Ni! — Method

element_fun_from_N_to_Ni!(expr_tree, vector_indices)

Change the indices of the variables of expr_tree given the order given by vector_indices. It it paired with get_elemental_variable to define the elemental element functions expression tree.

Example:

julia> element_fun_from_N_to_Ni!(:(x[4] + x[5]), [4,5])
:(x[1] + x[2])

ExpressionTreeForge.evaluate_expr_tree — Method

evaluate_expr_tree(expr_tree, x::AbstractVector)

Evaluate the expr_tree given the vector x.

Example:

julia> evaluate_expr_tree(:(x[1] + x[2]), ones(2))
2.

ExpressionTreeForge.evaluate_expr_tree_multiple_points — Method

evaluate_expr_tree_multiple_points(expression_tree::Any, x::AbstractVector{AbstractVector}})
evaluate_expr_tree_multiple_points(expression_tree::Any)

Evaluate the expression_tree with several points, represented as x.

ExpressionTreeForge.get_Ui — Method

Ui = get_Ui(indices::Vector{Int}, n::Int)

Create a sparse matrix Ui from indices computed by get_elemental_variable. Every index i (of indices) form a line of Ui corresponding to i-th Euclidian vector.

ExpressionTreeForge.get_bound — Method

(inf_bound, sup_bound) = get_bound(bound_tree::Bound_tree)

Retrieve the bounds of the root of bound_tree, the bounds of expression tree.

ExpressionTreeForge.get_convexity_status — Method

convexity_status = get_convexity_status(convexity_tree::M_convexity_detection.Convexity_tree)
convexity_status = get_convexity_status(complete_tree::M_implementation_complete_expr_tree.Complete_expr_tree)

Return the convexity status of either convexity_tree or complete_tree. The status can be:

constant

linear

strictly convex

strictly concave

unknown

ExpressionTreeForge.get_elemental_variable — Method

indices = get_elemental_variable(expr_tree)

Return the indices of the variable appearing in expr_tree. This function find the elemental variables from the expression tree of an element function.

Example:

julia> get_elemental_variable(:(x[1] + x[3]) )
[1, 3]
julia> get_elemental_variable(:(x[1]^2 + x[6] + x[2]) )
[1, 6, 2]

ExpressionTreeForge.get_function_of_evaluation — Method

eval_expression_tree = get_function_of_evaluation(expression_tree)

Return and evaluation function of expression_tree with better performance than the actual evaluate_expr_tree.

ExpressionTreeForge.get_type_tree — Method

type = get_type_tree(expr_tree)

Return the type of expr_tree. It can either be: constant, linear, quadratic, cubic or more.

Example:

julia> get_type_tree(:(5+4)) == constant
true
julia> get_type_tree(:(x[1])) == linear
true
julia> get_type_tree(:(x[1]* x[2])) == quadratic
true

ExpressionTreeForge.gradient_expr_tree_forward — Method

gradient = gradient_expr_tree_forward(expr_tree, x)

Evaluate the gradient of expr_tree at the point x with a forward automatic differentiation method.

Example:

julia> gradient_expr_tree_forward(:(x[1] + x[2]), rand(2))
[1.0 1.0]

ExpressionTreeForge.gradient_expr_tree_reverse — Method

gradient = gradient_expr_tree_reverse(expr_tree, x)

Evaluate the gradient of expr_tree at the point x with a reverse automatic differentiation method.

Example:

julia> gradient_expr_tree_reverse(:(x[1] + x[2]), rand(2))
[1.0 1.0]

ExpressionTreeForge.hessian_expr_tree — Method

hessian = hessian_expr_tree(expr_tree, x)

Evaluate the Hessian of expr_tree at the point x.

Example:

julia> hessian_expr_tree(:(x[1]^2 + x[2]), rand(2))
[2.0 0.0; 0.0 0.0]

ExpressionTreeForge.is_concave — Method

bool = is_concave(convexity_status)

Check if convexity_status is strictly concave.

ExpressionTreeForge.is_constant — Method

bool = is_constant(convexity_status)

Check if convexity_status is constant.

ExpressionTreeForge.is_constant — Method

bool = is_constant(type::Type_calculus_tree)

Check if type have the value constant from the enumerative type Type_calculus_tree.

ExpressionTreeForge.is_convex — Method

bool = is_convex(convexity_status)

Check if convexity_status is strictly convex.

ExpressionTreeForge.is_cubic — Method

bool = is_cubic(type::Type_calculus_tree)

Check if type have the value cubic from the enumerative type Type_calculus_tree.

ExpressionTreeForge.is_linear — Method

bool = is_linear(convexity_status)

Check if convexity_status is linear.

ExpressionTreeForge.is_linear — Method

bool = is_linear(type::Type_calculus_tree)

Check if type have the value linear from the enumerative type Type_calculus_tree.

ExpressionTreeForge.is_more — Method

bool = is_more(type::Type_calculus_tree)

Check if type have the value more from the enumerative type Type_calculus_tree.

ExpressionTreeForge.is_not_treated — Method

bool = is_not_treated(convexity_status)

Check if convexity_status is not_treated.

ExpressionTreeForge.is_quadratic — Method

bool = is_quadratic(type::Type_calculus_tree)

Check if type have the value quadratic from the enumerative type Type_calculus_tree.

ExpressionTreeForge.is_treated — Method

bool = is_treated(convexity_status)

Check if convexity_status is different of not_treated.

ExpressionTreeForge.is_unknown — Method

bool = is_unknown(convexity_status)

Check if convexity_status is unknown.

ExpressionTreeForge.linear_type — Method

linear = linear_type()

Return the value of linear from the enumerative type M_implementation_convexity_type.Convexity_type.

ExpressionTreeForge.print_tree — Method

print_tree(tree::AbstractTree)

Print a tree as long as it satisfies the interface M_interface_tree.

ExpressionTreeForge.set_bounds! — Method

set_bounds!(tree, bound_tree::Bound_tree)
set_bounds!(complete_tree::Complete_expr_tree)

Set the bounds of bound_tree by walking tree and propagate the computation from the leaves to the root. A Complete_expr_tree contains a precompiled bound_tree, and then can be use alone.

ExpressionTreeForge.set_convexity! — Method

set_convexity!(tree, convexity_tree, bound_tree)
set_convexity!(complete_tree)

Deduce from elementary rules the convexity status of tree nodes or complete_tree nodes. complete_tree integrate a bounds tree and can run alone the convexity detection whereas tree require the bound_tree (see create_bounds_tree) and convexity_tree (see create_convex_tree).

ExpressionTreeForge.transform_to_Expr — Method

expr = transform_to_Expr(expr_tree)

Transform expr_tree into an Expr. Warning: This function return an Expr with variables as MathOptInterface.VariableIndex. In order to get an standard Expr use transform_to_Expr_julia.

ExpressionTreeForge.transform_to_Expr_julia — Method

expr = transform_to_Expr_julia(expr_tree)

Transform expr_tree into an Expr.

ExpressionTreeForge.transform_to_expr_tree — Method

expr_tree = transform_to_expr_tree(expr)

Transform expr into an expr_tree::Type_expr_tree.

ExpressionTreeForge.unknown_type — Method

unknown = unknown_type()

Return the value of unknown from the enumerative type M_implementation_convexity_type.Convexity_type.

ExpressionTreeForge.M_abstract_expr_tree.AbstractExprTree — Type

Supertype of every expression tree.

ExpressionTreeForge.M_abstract_expr_tree.Bounds — Type

Bounds{T <: Number}

Structure representing the upper-bound and the lower-bound of a node.

ExpressionTreeForge.M_abstract_expr_tree.create_Expr — Method

expr = create_expr_tree(tree)

Create an Expr from tree. tree can be several types. For now, it supports Expr and ModelingToolkit.Operation, as well as the internal expression trees defined. The variables of the Expr use MathOptInterface variables.

ExpressionTreeForge.M_abstract_expr_tree.create_Expr2 — Method

expr = create_expr_tree(tree)

Create a julia Expr from tree. tree can be several types. For now, it supports Expr and ModelingToolkit.Operation, as well as the internal expression trees defined.

ExpressionTreeForge.M_abstract_expr_tree.create_empty_bounds — Method

bound = create_empty_bounds(type::DataType)

Create a Bounds structure of type.

ExpressionTreeForge.M_abstract_expr_tree.create_expr_tree — Method

expr_tree = create_expr_tree(tree)

Create an Type_expr_tree from tree. tree can be several types. For now, it supports Expr and ModelingToolkit.Operation, as well as the internal expression trees defined.

ExpressionTreeForge.M_abstract_expr_tree.get_bounds — Method

(inf_bound, sup_bound) = get_bounds(bound::Bounds{T}) where {T <: Number}

Retrieve both inf_bound and sup_bound from bound.

ExpressionTreeForge.M_abstract_expr_tree.set_bound! — Method

set_bound!(bound::Bounds{T}, inf_bound::T, sup_bound::T) where {T <: Number}

Set bound to inf_bound and sup_bound.

ExpressionTreeForge.M_trait_expr_tree.Is_expr_tree — Type

Type instantiated dynamically checking that an argument is an expression tree.

ExpressionTreeForge.M_trait_expr_tree.Is_not_expr_tree — Type

Type instantiated dynamically checking that an argument is not an expression tree.

ExpressionTreeForge.M_trait_expr_tree.expr_tree_equal — Function

bool = expr_tree_equal(expr_tree1, expr_tree2, eq::Atomic{Bool} = Atomic{Bool}(true))

Check if expr_tree1 and expr_tree2 valued the same, by checking recursively their respective nodes.

ExpressionTreeForge.M_trait_expr_tree.expr_tree_to_create — Method

expr_tree = expr_tree_to_create(tree1::AbstractExprTree, tree2::AbstractExprTree)

Return expr_tree, a expression tree of the type from tree2 with the values of tree1. It is supported for few expression-tree structure.

ExpressionTreeForge.M_trait_expr_tree.get_expr_children — Method

children = get_expr_children(tree::AbstractExprTree)

Return the children of the root of tree.

ExpressionTreeForge.M_trait_expr_tree.get_expr_node — Method

node = get_expr_node(tree::AbstractExprTree)

Return the node being the root of the tree.

ExpressionTreeForge.M_trait_expr_tree.get_real_node — Method

_get_real_node(tree::AbstractExprTree)

Return the value of a leaf in a suitable format for particular algorithm.

ExpressionTreeForge.M_trait_expr_tree.inverse_expr_tree — Method

minus_tree = inverse_expr_tree(tree::AbstractExprTree)

Apply a unary minus on tree.

ExpressionTreeForge.M_trait_expr_tree.is_expr_tree — Method

verifier_type = is_expr_tree(arg)

Return Is_expr_tree if arg is considered as an implementation of an expression tree, it returns Is_not_expr_tree otherwise.

ExpressionTreeForge.M_trait_expr_tree.transform_to_Expr — Method

expr = transform_to_Expr(expr_tree)

Return expr::Expr from the expression tree expr_tree.

ExpressionTreeForge.M_trait_expr_tree.transform_to_expr_tree — Method

expr_tree = _transform_to_expr_tree(tree::AbstractExprTree)

Return expr_tree::Type_expr_tree from tree. Type_expr_tree is the internal expression tree structure.

ExpressionTreeForge.M_interface_expr_tree._expr_tree_to_create — Method

expr_tree = _expr_tree_to_create(tree1::AbstractExprTree, tree2::AbstractExprTree)

Return expr_tree, a expression tree of the type from tree2 with the values of tree1. It is supported for few expression-tree structure.

ExpressionTreeForge.M_interface_expr_tree._get_expr_children — Method

children = _get_expr_children(tree::AbstractExprTree)

Return the children of the root of tree.

ExpressionTreeForge.M_interface_expr_tree._get_expr_node — Method

node = _get_expr_node(tree::AbstractExprTree)

Return the node being the root of the tree.

ExpressionTreeForge.M_interface_expr_tree._get_real_node — Method

_get_real_node(tree::AbstractExprTree)

Return the value of a leaf in a suitable format for particular algorithm.

ExpressionTreeForge.M_interface_expr_tree._inverse_expr_tree — Method

minus_tree = _inverse_expr_tree(tree::AbstractExprTree)

Apply a unary minus on tree.

ExpressionTreeForge.M_interface_expr_tree._transform_to_expr_tree — Method

expr_tree = _transform_to_expr_tree(tree::AbstractExprTree)

Return expr_tree::Type_expr_tree from tree. Type_expr_tree is the internal expression tree structure.

ExpressionTreeForge.M_implementation_expression_tree_Expr.Variable_counter — Type

Variable_counter

Associate the index of the symbol variable to their integer index.

ExpressionTreeForge.M_implementation_expr_tree.Type_expr_tree — Type

Type_expr_tree{T} <: AbstractTree

Basic implementation of an expression tree.Every expression tree supported must be able to returnTypeexprtree. ATypeexprtree` has fields:

field::Abstract_expr_node representing an operator, a constant or a variable;

children::Vector{Type_expr_tree{T}} a vector of children, each of them being a Type_expr_tree{T}.

ExpressionTreeForge.M_implementation_complete_expr_tree.Complete_expr_tree — Type

Complete_expr_tree{T} <: AbstractTree

Implementation of an expression tree. Complete_expr_tree is the same than Type_expr_tree with the additions in each node of a Bounds and Convexity_wrapper. A Complete_expr_tree has fields:

field::Complete_node{T} representing an operator, a constant or a variable alongide its bounds and its convexity status;

children::Vector{Complete_expr_tree{T}} a vector of children, each of them being a Complete_expr_tree{T}.

ExpressionTreeForge.M_implementation_complete_expr_tree.Complete_node — Type

Complete_node{T <: Number}

Gather in a single structure:

op::M_abstract_expr_node.Abstract_expr_node an arithmetic operator;

bounds::M_abstract_expr_tree.Bounds{T} the bounds of this operator depending its children;

convexity_status::M_implementation_convexity_type.Convexity_wrapper the convexity status of this operator depending its children.

ExpressionTreeForge.M_evaluation_expr_tree.gradient_expr_tree_forward — Method

gradient_expr_tree_forward(expr_tree, x)

Evaluate the gradient of the function represented by `expr_tree at the point x.

ExpressionTreeForge.M_evaluation_expr_tree.hessian_expr_tree — Method

hessian_expr_tree(expr_tree, x)

Compute the hessian matrix of the function represented by expr_tree at the point x.

ExpressionTreeForge.M_bound_propagations.Bound_tree — Type

Bound_tree

A tree where each node is a pair (bound_inf, bound_sup). Must be paired with an expression tree.

ExpressionTreeForge.M_bound_propagations.create_bounds_tree — Function

bound_tree = create_bounds_tree(tree)

Return a similar expression tree to tree, where each node has an undefined bounds.

ExpressionTreeForge.M_bound_propagations.get_bound — Method

(inf_bound, sup_bound) = get_bound(bound_tree::Bound_tree)

Retrieve the bounds of the root of bound_tree, the bounds of expression tree.

ExpressionTreeForge.M_bound_propagations.set_bounds! — Method

set_bounds!(tree, bound_tree::Bound_tree)
set_bounds!(complete_tree::Complete_expr_tree)

Set the bounds of bound_tree by walking tree and by propagating the computations from the leaves to the root. A Complete_expr_tree contains a precompiled bound_tree, and then can be use alone.

ExpressionTreeForge.M_convexity_detection.Convexity_tree — Type

Convexity_tree

A tree where each node is a convexity status. Must be paired with an expression tree.

ExpressionTreeForge.M_convexity_detection.concave_type — Method

concave = concave_type()

Return the value concave from the enumerative type M_implementation_convexity_type.Convexity_type.

ExpressionTreeForge.M_convexity_detection.constant_type — Method

constant = constant_type()

Return the value constant from the enumerative type M_implementation_convexity_type.Convexity_type.

ExpressionTreeForge.M_convexity_detection.convex_type — Method

convex = convex_type()

Return the value convex from the enumerative type M_implementation_convexity_type.Convexity_type.

ExpressionTreeForge.M_convexity_detection.create_convex_tree — Method

convex_tree = create_convex_tree(tree::Type_node)

Return a similar expression tree to tree, where each node has an undefined convexity status.

ExpressionTreeForge.M_convexity_detection.get_convexity_status — Method

convexitystatus = getconvexitystatus(convexitytree::Mconvexitydetection.Convexitytree) convexitystatus = getconvexitystatus(completetree::Mimplementationcompleteexprtree.Completeexpr_tree)

Return the convexity status of either convexity_tree or complete_tree. The status can be:

constant

linear

strictly convex

strictly concave

unknown

ExpressionTreeForge.M_convexity_detection.linear_type — Method

linear = linear_type()

Return the value linear from the enumerative type M_implementation_convexity_type.Convexity_type.

ExpressionTreeForge.M_convexity_detection.set_convexity! — Method

set_convexity!(tree, convexity_tree, bound_tree)
set_convexity!(complete_tree)

Deduce from elementary rules the convexity status of tree nodes or complete_tree nodes. complete_tree integrate a bounds tree and can run alone the convexity detection whereas tree require the bound_tree (see create_bounds_tree) and convexity_tree (see create_convex_tree).

ExpressionTreeForge.M_convexity_detection.unknown_type — Method

unknown = unknown_type()

Return the value unknown from the enumerative type M_implementation_convexity_type.Convexity_type.

ExpressionTreeForge.algo_expr_tree.N_to_Ni — Method

dic = N_to_Ni(elemental_var::Vector{Int})

Return a dictionnary informing the index changes of an element expression tree. If element_var = [4,5] then dic == Dict([4=>1, 5=>2]).

ExpressionTreeForge.algo_expr_tree.cast_type_of_constant — Method

cast_type_of_constant(expr_tree, type::DataType)

Cast to type the constants of expr_tree.

ExpressionTreeForge.algo_expr_tree.delete_imbricated_plus — Method

separated_terms = delete_imbricated_plus(expr_tree)

Divide the expression tree as a terms of a sum if possible. It returns a vector where each component is a subexpression tree of expr_tree.

Example:

julia> delete_imbricated_plus(:(x[1] + x[2] + x[3]*x[4] ) )
3-element Vector{Expr}:
 :(x[1])
 :(x[2])
 :(x[3] * x[4])

ExpressionTreeForge.algo_expr_tree.element_fun_from_N_to_Ni! — Method

element_fun_from_N_to_Ni!(expr_tree, vector_indices)

Change the indices of the variables of expr_tree given the order given by vector_indices. It it paired with get_elemental_variable to define the elemental element functions expression tree.

Example:

julia> element_fun_from_N_to_Ni!(:(x[4] + x[5]), [4,5])
:(x[1] + x[2])

ExpressionTreeForge.algo_expr_tree.get_Ui — Method

Ui = get_Ui(indices::Vector{Int}, n::Int)

Create a sparse matrix Ui from indices computed by get_elemental_variable. Every index i (of indices) form a line of Ui corresponding to i-th Euclidian vector.

ExpressionTreeForge.algo_expr_tree.get_elemental_variable — Method

indices = get_elemental_variable(expr_tree)

Return the indices of the variable appearing in expr_tree. This function find the elemental variables from the expression tree of an element function.

Example:

julia> get_elemental_variable(:(x[1] + x[3]) )
[1, 3]
julia> get_elemental_variable(:(x[1]^2 + x[6] + x[2]) )
[1, 6, 2]

ExpressionTreeForge.algo_expr_tree.get_type_tree — Method

type = get_type_tree(expr_tree)

Return the type of expr_tree. It can either be: constant, linear, quadratic, cubic or more.

Example:

julia> get_type_tree(:(5+4)) == constant
true
julia> get_type_tree(:(x[1])) == linear
true
julia> get_type_tree(:(x[1]* x[2])) == quadratic
true

ExpressionTreeForge.M_trait_expr_node.Type_expr_node — Type

Type instantiated dynamically checking that an argument is an expression node.

ExpressionTreeForge.M_trait_expr_node.Type_not_expr_node — Type

Type instantiated dynamically checking that an argument is a not an expression node.

ExpressionTreeForge.M_trait_expr_node.cast_constant! — Method

bool = cast_constant!(node::Abstract_expr_node, type::Datatype)

Rewrite the Numbers composing node as type.

ExpressionTreeForge.M_trait_expr_node.change_from_N_to_Ni! — Method

change_from_N_to_Ni!(node::Abstract_expr_node, dic_new_indices::Dict{Int, Int})

Change the index of the variables following the index given by dic_new_indices.

ExpressionTreeForge.M_trait_expr_node.evaluate_node! — Method

evaluate_node!(node::Abstract_expr_node, value_children::AbstractVector{T}, ref::MyRef{T}) where T<:Number

Evaluate node depending value_children and store the result in ref.

ExpressionTreeForge.M_trait_expr_node.evaluate_node — Method

value = _evaluate_node(node::Abstract_expr_node, value_children::AbstractVector{T}) where T<:Number

Evaluate node depending value_children.

ExpressionTreeForge.M_trait_expr_node.get_type_node — Method

type = get_type_node(node::Abstract_expr_node, children_types::AbstractVector{Type_expr_basic})

Return the type of node given the type of its children children_types. The types avaible areconstant, linear, quadratic, cubic or more.

ExpressionTreeForge.M_trait_expr_node.get_var_index — Method

index = get_var_index(variable::Abstract_expr_node)

Return the index of the variable.

ExpressionTreeForge.M_trait_expr_node.node_bound — Method

(lowerbound, upper_bound) = node_bound(node::Abstract_expr_node, children_bounds::AbstractVector{Tuple{T, T}}, type::DataType)

Return the bounds of node given the children_bounds.

ExpressionTreeForge.M_trait_expr_node.node_convexity — Method

bool = _node_is_operator(node::Abstract_expr_node, children_convex_status::AbstractVector{M_implementation_convexity_type.Convexity_type})

Return the convexity status of node given the children_convex_status.

ExpressionTreeForge.M_trait_expr_node.node_is_constant — Method

bool = node_is_constant(node::Abstract_expr_node)

Check if node is a constant.

ExpressionTreeForge.M_trait_expr_node.node_is_cos — Method

bool = node_is_cos(node::Abstract_expr_node)

Check if node is a cos operator.

ExpressionTreeForge.M_trait_expr_node.node_is_minus — Method

bool = node_is_minus(node::Abstract_expr_node)

Check if node is a - operator or not.

ExpressionTreeForge.M_trait_expr_node.node_is_operator — Method

bool = node_is_operator(node::Abstract_expr_node)

Check if node is an operator. It could be also a variable or a constant.

ExpressionTreeForge.M_trait_expr_node.node_is_plus — Method

bool = node_is_plus(node::Abstract_expr_node)

Check if node is a + operator or not.

ExpressionTreeForge.M_trait_expr_node.node_is_power — Method

bool = node_is_power(node::Abstract_expr_node)

Check if node is a ^ operator or not.

ExpressionTreeForge.M_trait_expr_node.node_is_sin — Method

bool = node_is_sin(node::Abstract_expr_node)

Check if node is a sin operator.

ExpressionTreeForge.M_trait_expr_node.node_is_tan — Method

bool = node_is_tan(node::Abstract_expr_node)

Check if node is a tan operator.

ExpressionTreeForge.M_trait_expr_node.node_is_times — Method

bool = _node_is_times(node::Abstract_expr_node)

Check if node is a * operator or not.

ExpressionTreeForge.M_trait_expr_node.node_is_variable — Method

bool = _node_is_variable(node::Abstract_expr_node)

Check if node is a variable. It could be also an operator or a constant.

ExpressionTreeForge.M_trait_expr_node.node_to_Expr — Method

bool = node_to_Expr(node::Abstract_expr_node)

Return the information required to build later on a julia Expr. An operator return the operator symbol (ex: + -> :+). A variable return a variable as a MathOptInterface variable. A constant return its value.

ExpressionTreeForge.M_trait_expr_node.node_to_Expr2 — Method

bool = node_to_Expr2(node::Abstract_expr_node)

Return the information required to build later on a julia Expr. An operator return the operator symbol (ex: + -> :+). A variable return a variable. A constant return its value.

ExpressionTreeForge.M_interface_expr_node._cast_constant! — Method

bool = _cast_constant!(node::Abstract_expr_node, type::Datatype)

Rewrite the Numbers composing node as type.

ExpressionTreeForge.M_interface_expr_node._change_from_N_to_Ni! — Method

_change_from_N_to_Ni!(node::Abstract_expr_node, dic_new_indices::Dict{Int, Int})

Change the index of the variables following the index given by dic_new_indices.

ExpressionTreeForge.M_interface_expr_node._evaluate_node! — Method

_evaluate_node!(node::Abstract_expr_node, value_children::AbstractVector{T}, ref::MyRef{T}) where T<:Number

Evaluate node depending value_children and store the result in ref.

ExpressionTreeForge.M_interface_expr_node._evaluate_node — Method

value = _evaluate_node(node::Abstract_expr_node, value_children::AbstractVector{T}) where T<:Number

Evaluate node depending value_children.

ExpressionTreeForge.M_interface_expr_node._get_type_node — Method

type = _get_type_node(node::Abstract_expr_node, children_types::AbstractVector{Type_expr_basic})

Return the type of node given the type of its children children_types. The types available are constant, linear, quadratic, cubic or more.

ExpressionTreeForge.M_interface_expr_node._get_var_index — Method

index = _get_var_index(variable::Abstract_expr_node)

Return the index of the variable.

ExpressionTreeForge.M_interface_expr_node._node_bound — Method

(lowerbound, upper_bound) = _node_bound(node::Abstract_expr_node, children_bounds::AbstractVector{Tuple{T, T}}, type::DataType)

Return the bounds of node given the children_bounds.

ExpressionTreeForge.M_interface_expr_node._node_convexity — Method

bool = _node_is_operator(node::Abstract_expr_node, children_convex_status::AbstractVector{M_implementation_convexity_type.Convexity_type})

Return the convexity status of node given the children_convex_status.

ExpressionTreeForge.M_interface_expr_node._node_is_constant — Method

bool = _node_is_constant(node::Abstract_expr_node)

Check if node is a constant.

ExpressionTreeForge.M_interface_expr_node._node_is_cos — Method

bool = _node_is_cos(node::Abstract_expr_node)

Check if node is a cos operator.

ExpressionTreeForge.M_interface_expr_node._node_is_minus — Method

bool = _node_is_minus(node::Abstract_expr_node)

Check if node is a - operator or not.

ExpressionTreeForge.M_interface_expr_node._node_is_operator — Method

bool = _node_is_operator(node::Abstract_expr_node)

Check if node is an operator. It could be also a variable or a constant.

ExpressionTreeForge.M_interface_expr_node._node_is_plus — Method

bool = _node_is_plus(node::Abstract_expr_node)

Check if node is a + operator or not.

ExpressionTreeForge.M_interface_expr_node._node_is_power — Method

bool = _node_is_power(node::Abstract_expr_node)

Check if node is a ^ operator or not.

ExpressionTreeForge.M_interface_expr_node._node_is_sin — Method

bool = _node_is_sin(node::Abstract_expr_node)

Check if node is a sin operator.

ExpressionTreeForge.M_interface_expr_node._node_is_tan — Method

bool = _node_is_tan(node::Abstract_expr_node)

Check if node is a tan operator.

ExpressionTreeForge.M_interface_expr_node._node_is_times — Method

bool = _node_is_times(node::Abstract_expr_node)

Check if node is a * operator or not.

ExpressionTreeForge.M_interface_expr_node._node_is_variable — Method

bool = _node_is_variable(node::Abstract_expr_node)

Check if node is a variable. It could be also an operator or a constant.

ExpressionTreeForge.M_interface_expr_node._node_to_Expr — Method

bool = _node_to_Expr(node::Abstract_expr_node)

Return the information required to build later on a julia Expr. An operator return the operator symbol (ex: + -> :+). A variable return a variable as a MathOptInterface variable. A constant return its value.

ExpressionTreeForge.M_interface_expr_node._node_to_Expr2 — Method

bool = _node_to_Expr2(node::Abstract_expr_node)

Return the information required to build later on a julia Expr. An operator return the operator symbol (ex: + -> :+). A variable return a variable. A constant return its value.

ExpressionTreeForge.M_abstract_expr_node.Abstract_expr_node — Type

Supertype of every node.

ExpressionTreeForge.M_abstract_expr_node.MyRef — Type

MyRef{Y <: Number}

Sort of pointer. Contain the field:

value which points to the desired structure.

ExpressionTreeForge.M_abstract_expr_node.create_new_vector_myRef — Function

vector_myref = create_new_vector_myRef(n::Int, type::DataType = Float64)

Create a vector of n myRef{type} components.

ExpressionTreeForge.M_abstract_expr_node.create_node_expr — Method

create_node_expr(arg::UnionAll)

Create a node from arg. See the different implementation in the src/node_expr_tree/impl_operators.jl.

ExpressionTreeForge.M_abstract_expr_node.create_undef_array_myRef — Function

my_ref_matrix = create_undef_array_myRef(n::Int, m::Int, type::DataType = Float64)

Create a n×m array composed of myRef{type} components.

ExpressionTreeForge.M_abstract_expr_node.create_vector_of_vector_myRef — Function

vector_vector_myref = create_undef_array_myRef(l::Int, c::Int, type::DataType = Float64)

Create a vector of size l, each component is a Vector{myRef{type}} of c components.

ExpressionTreeForge.M_abstract_expr_node.equalize_vec_vec_myRef! — Method

equalize_vec_vec_myRef!(a::Vector{Vector{MyRef{T}}}, b::Vector{Vector{MyRef{T}}}) where {T <: Number}

Set the references of b to those of a.

ExpressionTreeForge.M_abstract_expr_node.get_myRef — Method

value = get_myRef(reference::MyRef{Y}) where {Y <: Number}

Get the value of the reference.

ExpressionTreeForge.M_abstract_expr_node.new_ref — Method

myref = new_ref(value::Y) where {Y <: Number}
myref = new_ref(type::DataType)

Create a new reference myRef from value or type.

ExpressionTreeForge.M_abstract_expr_node.set_myRef! — Method

set_myRef!(reference::MyRef{Y}, value::Y) where {Y <: Number}

Set the reference to value.

ExpressionTreeForge.M_trait_tree.Type_not_trait_tree — Type

Type instantiated dynamically checking that an argument is not a tree.

ExpressionTreeForge.M_trait_tree.Type_trait_tree — Type

Type instantiated dynamically checking that an argument is a tree.

ExpressionTreeForge.M_interface_tree._get_children — Method

children = _get_children(tree, trait_tree)

Get the children from the current node as part of tree, if trait_tree::Type_trait_tree..

ExpressionTreeForge.M_interface_tree._get_node — Method

node = _get_node(tree, trait_tree)

Get the current node as a part of tree, if trait_tree::Type_trait_tree.

ExpressionTreeForge.M_trait_tree.get_children — Method

children = get_children(tree)

Get the children from the current node as part of tree.

ExpressionTreeForge.M_trait_tree.get_node — Method

node = get_node(tree)

Get the current node as a part of tree.

ExpressionTreeForge.M_trait_tree.is_type_trait_tree — Method

Type_trait_tree = is_type_trait_tree(a::AbstractTree)
Type_trait_tree = is_type_trait_tree(a::Expr)
Type_trait_tree = is_type_trait_tree(a::Number)
Type_not_trait_tree = is_type_trait_tree(a::Any)

Return a Type_trait_tree only for AbstractTree, Expr or Number (a leaf of a tree). In the other cases it returns Type_not_trait_tree.

ExpressionTreeForge.algo_tree.printer_tree — Function

print_tree(tree::AbstractTree)

Print a tree as long as it satisfies the interface M_interface_tree.

ExpressionTreeForge.M_interface_tree._get_children — Method

children = _get_children(tree)

Get the children from the current node as part of tree.

ExpressionTreeForge.M_interface_tree._get_node — Method

node = _get_node(tree)

Get the current node as a part of tree.

ExpressionTreeForge.M_implementation_tree.Type_node — Type

Type_node{T} <: AbstractTree

Basic implementation of a tree. A Type_node has fields:

field gathering the informations about the current node;

children a vector of children, each of them being a Type_node.

ExpressionTreeForge.M_implementation_tree.equal_tree — Method

bool = equal_tree(tree1::Type_node{T}, tree2::Type_node{T})

Check recursively if every node is the same for both tree1 and tree2

ExpressionTreeForge.M_abstract_tree.AbstractTree — Type

Supertype of every tree.

ExpressionTreeForge.M_abstract_tree.create_tree — Method

tree = create_tree(field::T, children::Vector{Type_node{T}}) where {T}
tree = create_tree(field::T, children::Array{Any, 1}) where {T}
tree = create_tree(ex::Expr)

Create a tree of type Type_node from a field for the current node and its children or from an Expr.

ExpressionTreeForge.M_trait_type_expr.Type_not_type_expr — Type

Type instantiated dynamically that checks if an argument is not a Type_expr_basic

ExpressionTreeForge.M_trait_type_expr.Type_type_expr — Type

Type instantiated dynamically that checks if an argument is a Type_expr_basic

ExpressionTreeForge.M_trait_type_expr.is_constant — Method

bool = _is_constant(type)

Return true if type is a constant type.

ExpressionTreeForge.M_trait_type_expr.is_cubic — Method

bool = _is_cubic(type)

Return true if type is a cubic type.

ExpressionTreeForge.M_trait_type_expr.is_linear — Method

bool = _is_linear(type)

Return true if type is a linear type.

ExpressionTreeForge.M_trait_type_expr.is_more — Method

bool = _is_more(type)

Return true if type is more non linear than a cubic.

ExpressionTreeForge.M_trait_type_expr.is_quadratic — Method

bool = _is_quadratic(type)

Return true if type is a quadratic type.

ExpressionTreeForge.M_trait_type_expr.is_trait_type_expr — Method

type = is_trait_type_expr(arg)

Return type::Type_type_expr if arg is an Type_expr_basic otherwise is return type::Type_not_type_expr

ExpressionTreeForge.M_trait_type_expr.type_power — Method

result_type = type_power(index, type)

Return the type resulting of type to power index.

ExpressionTreeForge.M_trait_type_expr.type_product — Method

result_type = type_product(type1, type2)

Return the type resulting of a product between type1 and type2.

ExpressionTreeForge.M_interface_type_expr._is_constant — Method

bool = _is_constant(type)

Return true if type is a constant type.

ExpressionTreeForge.M_interface_type_expr._is_cubic — Method

bool = _is_cubic(type)

Return true if type is a cubic type.

ExpressionTreeForge.M_interface_type_expr._is_linear — Method

bool = _is_linear(type)

Return true if type is a linear type.

ExpressionTreeForge.M_interface_type_expr._is_more — Method

bool = _is_more(type)

Return true if type is more non linear than a cubic type.

ExpressionTreeForge.M_interface_type_expr._is_quadratic — Method

bool = _is_quadratic(type)

Return true if type is a quadratic type.

ExpressionTreeForge.M_interface_type_expr._type_power — Method

result_type = _type_power(index, type)

Return the type resulting of type to power index.

ExpressionTreeForge.M_interface_type_expr._type_product — Method

result_type = _type_product(type1, type2)

Return the type resulting of a product between type1 and type2.

ExpressionTreeForge.M_implementation_type_expr.return_constant — Method

constant = return_constant()

Return constantype::Type_expr_basic.

ExpressionTreeForge.M_implementation_type_expr.return_cubic — Method

cubic = return_cubic()

Return cubic::Type_expr_basic.

ExpressionTreeForge.M_implementation_type_expr.return_linear — Method

linear = return_linear()

Return linear::Type_expr_basic.

ExpressionTreeForge.M_implementation_type_expr.return_more — Method

more = return_more()

Return more::Type_expr_basic.

ExpressionTreeForge.M_implementation_type_expr.return_quadratic — Method

quadratic = return_quadratic()

Return quadratic::Type_expr_basic.

ExpressionTreeForge.M_interface_type_expr._is_constant — Method

bool = _is_constant(type::Type_expr_basic)

Check if type equals constant.

ExpressionTreeForge.M_interface_type_expr._is_cubic — Method

bool = _is_cubic(type::Type_expr_basic)

Check if type equals cubic.

ExpressionTreeForge.M_interface_type_expr._is_linear — Method

bool = _is_linear(type::Type_expr_basic)

Check if type equals linear.

ExpressionTreeForge.M_interface_type_expr._is_more — Method

bool = _is_more(type::Type_expr_basic)

Check if type equals more.

ExpressionTreeForge.M_interface_type_expr._is_quadratic — Method

bool = _is_quadratic(type::Type_expr_basic)

Check if type equals quadratic.

ExpressionTreeForge.M_interface_type_expr._type_power — Method

result_type = _type_power(index_power::Number, b::Type_expr_basic)

Return result_type::Type_expr_basic, resulting of b^(index).

ExpressionTreeForge.M_interface_type_expr._type_product — Method

result_type = _type_product(a::Type_expr_basic, b::Type_expr_basic)

Return result_type::Type_expr_basic, the type resulting of the product a*b.

ExpressionTreeForge.M_implementation_convexity_type.Convexity_type — Type

Convexity_type

Type representing wether an expression is not_treated, constant, linear, convex, concave or unknown.

ExpressionTreeForge.M_implementation_convexity_type.Convexity_wrapper — Type

Convexity_wrapper

Wrapper around Convexity_type.

ExpressionTreeForge.M_implementation_convexity_type.concave_type — Method

concave = concave_type()

Return a concave::Convexity_type.

ExpressionTreeForge.M_implementation_convexity_type.concave_wrapper — Method

concave_wrapper = concave_wrapper()

Return a Convexity_wrapper with value concave.

ExpressionTreeForge.M_implementation_convexity_type.constant_type — Method

constant = constant_type()

Return a constant::Convexity_type.

ExpressionTreeForge.M_implementation_convexity_type.constant_wrapper — Method

constant_wrapper = constant_wrapper()

Return a Convexity_wrapper with value constant.

ExpressionTreeForge.M_implementation_convexity_type.convex_type — Method

convex = convex_type()

Return a convex::Convexity_type.

ExpressionTreeForge.M_implementation_convexity_type.convex_wrapper — Method

convex_wrapper = convex_wrapper()

Return a Convexity_wrapper with value convex.

ExpressionTreeForge.M_implementation_convexity_type.get_convexity_wrapper — Method

convexity_status = get_convexity_wrapper(convexity::Convexity_wrapper)

Return the convexity status of convexity.

ExpressionTreeForge.M_implementation_convexity_type.init_conv_status — Method

untreated_wrapper = init_conv_status()

Return a Convexity_wrapper with value not_treated.

ExpressionTreeForge.M_implementation_convexity_type.is_concave — Method

bool = is_concave(status::Convexity_type)
bool = is_concave(convexity::Convexity_wrapper)

Check if status or convexity.status equals concave.

ExpressionTreeForge.M_implementation_convexity_type.is_constant — Method

bool = is_constant(status::Convexity_type)
bool = is_constant(convexity::Convexity_wrapper)

Check if status or convexity.status equals constant.

ExpressionTreeForge.M_implementation_convexity_type.is_convex — Method

bool = is_convex(status::Convexity_type)
bool = is_convex(convexity::Convexity_wrapper)

Check if status or convexity.status equals convex.

ExpressionTreeForge.M_implementation_convexity_type.is_linear — Method

bool = is_linear(status::Convexity_type)
bool = is_linear(convexity::Convexity_wrapper)

Check if status or convexity.status equals linear.

ExpressionTreeForge.M_implementation_convexity_type.is_not_treated — Method

bool = is_not_treated(status::Convexity_type)
bool = is_not_treated(convexity::Convexity_wrapper)

Check if status or convexity.status equals not_treated.

ExpressionTreeForge.M_implementation_convexity_type.is_treated — Method

bool = is_treated(status::Convexity_type)
bool = is_treated(convexity::Convexity_wrapper)

Check if status or convexity.status equals not_treated.

ExpressionTreeForge.M_implementation_convexity_type.is_unknown — Method

bool = is_unknown(status::Convexity_type)
bool = is_unknown(convexity::Convexity_wrapper)

Check if status or convexity.status equals unknown.

ExpressionTreeForge.M_implementation_convexity_type.linear_type — Method

linear = linear_type()

Return a linear::Convexity_type.

ExpressionTreeForge.M_implementation_convexity_type.linear_wrapper — Method

linear_wrapper = linear_wrapper()

Return a Convexity_wrapper with value linear.

ExpressionTreeForge.M_implementation_convexity_type.not_treated_type — Method

not_treated = not_treated_type()

Return a not_treated::Convexity_type.

ExpressionTreeForge.M_implementation_convexity_type.set_convexity_wrapper! — Method

set_convexity_wrapper!(convexity::Convexity_wrapper, t::Convexity_type)

Set the convexity status of convexity to type.

ExpressionTreeForge.M_implementation_convexity_type.unknown_type — Method

unknown = unknown_type()

Return a unknown::Convexity_type.

ExpressionTreeForge.M_implementation_convexity_type.unknown_wrapper — Method

unknown_wrapper = unknown_wrapper()

Return a Convexity_wrapper with value unknown.